- fixed a few more leftover Clang warnings.

+++ src/common/objects/dobjgc.cpp
@@ -721,7 +721,7 @@ FAveragizer::FAveragizer()
 void FAveragizer::AddAlloc(size_t alloc)
 {
 	NewestPos = (NewestPos + 1) & (HistorySize - 1);
-	if (TotalCount < HistorySize)
+	if (TotalCount < (int)HistorySize)
 	{
 		TotalCount++;
 	}
